warmup

+++ main.cpp
@@ -31,7 +31,7 @@ int main()
             break;
 
          case '%':
-            cout << num1 << " % " << num2 << " = " << num1 % num2 << '\n';
+            cout << num1 << " % " << num2 << " = " << (int)num1 % (int)num2 << '\n';
             break;
 
         default:
